requirements.txt
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0
Django~=5.0.3
djangorestframework~=3.14.0